Create updateDisplay function
As the user clicks a number, the number should appear on the
display. The updateDisplay function accomplished this. The number
clicked may either be another digit of the current number or a
new number. The append argument accounts for this. The new number
is appended to the display if true and the current text content is
replaced with the new number if false.