Add event listeners for operator buttons
When an operator is clicked, the operator needs to be updated
appropriately. If there is no operator yet, or if there is an
operator but no second operand, then the value of the operator
is updated. Else, the previous expression is evaluated and then
stored in operand1 and the new operator is updated.