Add event listener to equals button
When the equals button is clicked, the current expression should
be evaluated. This only happens if there is at least one operand
and an operator. If there is one operand, then the expression is
evaluated with that operand as both operands in the expression.
The event listener for the equals button evaluates the expression
as described above when it is clicked and updates the display.