Add event listener to all clear button
When the all clear (AC) button is clicked, the display needs to
be cleared and the calculator needs to be ready for a new
expression. The event listener for the all clear button resets
all the relevant variables to an empty string and updates the
display to once again be empty.