Update operate function to prevent 0 division
The user should be shown an error message when trying to divide
a number by 0. The operate function has been updated to return
an empty string if the second operand is 0. The event listeners
for the buttons handle this empty string accordingly and display
"ERROR" instead of a result. This prevents users from trying to
do further operations with NaN.