Prevent numbers from overflowing on display
When the user types many numbers in a row, the number may become
too large and overflow on the display. Additionally, if the number
provided by the user is larger than the maximum safe integer in
JavaScript, then the result will not be accurate. For the sake of
this calculator, a number is limited to 10 digits. If the user tries
to type a number longer than 10 digits, then neither the operand
nor the display will be updated.