Fix behavior after expression is evaluated with =
When the user clicks the = button to evaluate an expression, the
result should be updated on the display. If the user then clicks
another number, a new expression is started and the old return
value is discarded. However, if the user presses an operator button,
the result of the = can be used for further expressions. This was
not the case before, but has been resolved.