Fix behavior when division by zero
When the user divides by zero, an error message should be displayed
and the user should only be able to start a new expression from
there. Before, there was an error that allowed the user to apply
operators to an error result, which would result in NaN outputs.
This has been fixed; when an error message is displayed, the only
buttons that will respond to input are the number buttons so that
the user can start a new expression.