Support operations with decimal values
The user should be able to type a decimal value, but only be
allowed to type one decimal point per operand. Before, the decimal
button was not operational. Now, it has been given an event
listener that appends a decimal point to the current operand
when clicked. If there is already a decimal point in the current
operand, then it is not appended.