Add functionality to change sign of number
When typing in a number, the user may want to make that number
negative before running the operation. Now, the +/- button has
this functionality. In its event listener, if the number is
positive, it becomes negative, and if it is already negative, it
becomes positive. The button only works if there is already a
number on the display.