Add necessary zeros to display
When the calculator is first loaded or after it has been cleared,
the display should show zero and the zero should be replaced by
whatever the user types. Additionally, if the user types a decimal
before having typed any numbers, that decimal should begin with a
0 (e.g. 0.9 instead of just displaying .9). These two features
have both been implemented.