Add support for percentage button
When the user clicks the percentage button, it should apply to the
current operand, dividing it by 100. However, if there is one
operand and an operator, but no second operand, and the user then
clicks the percentage button, the value of the second operand
becomes the first operand * (first operand / 100). This has now
been implemented in the event listener of the percentage button.