Change color of operator button when clicked
Since the operator does not appear on the display, only the
operands and results of expressions appear on the display, it
would be useful for the user to have a visual indicator of
which operator is currently being used. In order to give a
visual indicator, the color of the operator now changes to
orange when clicked in a valid manner and changes back to
blue when no longer active.